django_web/requirements.txt
argcomplete==1.9.3
awscli==1.16.179
boto3==1.9.158
botocore==1.12.169
certifi==2019.3.9
cfn-flip==1.2.0
chardet==3.0.4
Click==7.0
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.2
django-bootstrap-form==3.4
django-environ==0.4.5
django-extensions==2.1.9
django-filter==2.1.0
django-rest-swagger==2.2.0
django-tables2==2.0.6
djangorestframework==3.9.4
docutils==0.14
durationpy==0.5
future==0.16.0
gevent==1.4.0
greenlet==0.4.15
gunicorn==19.9.0
hjson==3.0.1
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
jmespath==0.9.3
kappa==0.6.0
lambda-packages==0.20.0
MarkupSafe==1.1.1
openapi-codec==1.3.2
placebo==0.9.0
psycopg2-binary==2.8.3
pyasn1==0.4.5
python-dateutil==2.6.1
python-slugify==1.2.4
pytz==2019.1
PyYAML==3.13
requests==2.22.0
rsa==3.4.2
s3transfer==0.2.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
toml==0.10.0
tqdm==4.19.1
troposphere==2.4.7
Unidecode==1.0.23
uritemplate==3.0.0
urllib3==1.25.3
Werkzeug==0.15.4
whitenoise==4.1.2
wsgi-request-logger==0.4.6
zappa==0.48.2
